+++ assets/javascripts/kitten/components/typography/text/index.js
@@ -70,6 +70,9 @@ export class Text extends Component {
 }
 
 Text.propTypes = {
+  /**
+    Available colors:
+  */
   color: PropTypes.oneOf([
     'font1',
     'font2',
@@ -78,12 +81,33 @@ Text.propTypes = {
     'error',
     'valid',
   ]),
+  /**
+    If `tag="a"`, show underline.
+  */
   decoration: PropTypes.oneOf(['underline', 'none']),
-  lineHeight: PropTypes.oneOf(['normal', '1', '1.3']),
+  /**
+    `font-feature-settings: 'tnum'` enables tabular (monospace) numerals.
+  **/
   setting: PropTypes.oneOf(['tnum']),
+  /**
+    `line-height: normal` correspond approximately to 1.2.
+  */
+  lineHeight: PropTypes.oneOf(['normal', '1', '1.3']),
+  /**
+    Available sizes:
+  */
   size: PropTypes.oneOf(['huge', 'big', 'default', 'tiny', 'micro', 'nano']),
+  /**
+    Available font styles (`normal` or `italic`):
+  */
   fontStyle: PropTypes.oneOf(['normal', 'italic']),
+  /**
+    `text-transform: uppercase`
+  */
   transform: PropTypes.oneOf(['uppercase']),
+  /**
+    Available font weights (`light`, `regular` or `italic`):
+  */
   weight: PropTypes.oneOf(['light', 'regular', 'bold']),
 }